refactor: jackson 모듈 및 테스트 파일 패키지 이동
이동된 패키지
- common/domain/vo/configuration/jackson => common/configuration/jackson
